se arreglaron componentes
